+++ apps/wizard/app_pages/dataset_preview/dependency_graph.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, List, cast
 import streamlit as st
 from streamlit_agraph import Config, Edge, Node, agraph
 
-from etl.dag_utils import load_dag
+from etl.dag_helpers import load_dag
 from etl.steps import extract_step_attributes, filter_to_subgraph
 
 COLORS = {
