+++ apps/wizard/etl_steps/forms.py
@@ -14,7 +14,7 @@ from typing_extensions import Self
 from apps.utils.files import generate_step_to_channel
 from apps.wizard.etl_steps.utils import ADD_DAG_OPTIONS, COOKIE_STEPS, SNAPSHOT_SCHEMA, remove_playground_notebook
 from apps.wizard.utils import clean_empty_dict
-from etl.dag_utils import write_to_dag_file
+from etl.dag_helpers import write_to_dag_file
 from etl.files import ruamel_dump
 from etl.paths import DAG_DIR
