+++ etl/command.py
@@ -25,7 +25,7 @@ import structlog
 from ipdb import launch_ipdb_on_exception
 
 from etl import config, files, paths
-from etl.dag_utils import load_dag
+from etl.dag_helpers import load_dag
 from etl.steps import (
     DAG,
     DataStep,
