+++ etl/datadiff.py
@@ -20,7 +20,7 @@ from rich.panel import Panel
 from rich.syntax import Syntax
 from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_exponential
 
-from etl.dag_utils import load_dag
+from etl.dag_helpers import load_dag
 from etl.files import yaml_dump
 from etl.tempcompare import series_equals
