+++ etl/to_graphviz.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 import rich_click as click
 
-from etl.dag_utils import load_dag
+from etl.dag_helpers import load_dag
 from etl.steps import filter_to_subgraph
